Ignore password expiry if account does not have a password
If a new user is created, this user will not have a password by default
so a login is not possible. This applies to some machine users as well
(e.g. postgres). Instead of forcing the user to define a password aging,
when there is no need to even set a password, we now skip such users.