Ease link lists structure
If we want to insert to the end of a linked list we have to have a
pointer that points to the end (if we do not want to traverse the whole
list to find the end node).

But instead we can simply add to the beginning of the list. This makes
the data structure a lot easier.