Use binary search for sorted lists
If we are searching for files with an invalid owner or group, we do not
want run getpwuid and getgrgid methods for every file.

As a result we already cache the known uids and gids in an array that
can be checked for each file.

Since these arrays are already numerically sorted we can simply use a
binary search to make this even faster.