Allow --fast to skip expensive tests
Before this commit we did not allow any command line arguments. We now
allow to specify `--fast` which does currently only impact the file
collector. If we specify `--fast` we enable the collection flag
`COLLECTOR_FAST` and then skip the checking of the whole directory tree
to find unowned files or directories which are world-writeable.

This greatly speeds up a hardening check. The skipped tests will
currently count as `passed` but will show up as `SKIP` in the final
report.