Refactor user/group cache
Before this change we only had a static cache of uids in `util.c` and we
created temporary caches of usernames and ids in `user.c`.

We now create all caches in one go in `util.c` and use these caches in
`user.c` We also change the way we detect duplicate uids and names.

If we have multiple duplicates, e.g. usera, userb and userc have the
same uid 1000, we generated two findings before this change:
- userb has the same uid as usera
- userc has the same uid as userb

We now generate 3 findings:
- found 3 users with uid 1000: usera
- found 3 users with uid 1000: userb
- found 3 users with uid 1000: userc