Cache all user details
Right now we only build a cache of users with name and uid set. Right
now these are the only attributes that we care about because the cache
is mostly used to answer the question if a user with a certain uid is
present on the system or if we have users with duplicate uids or names.

We now cache all attributes so we can possibly use the cache in other
situations as well.