Do not call getpwuid in a setpwent/endpwent loop
In some situations getpwuid seems to reset the pointer to the current
user of a setpwent/getpwent/endpwent loop. This can cause the loop
to run forever.

This has happened in users.c where the loop over every user and check
the owner of the homedirectory. If we see that the uid of the
homedirectory does not match, we use getpwuid to get the owner of the
homedirectory to create a finding like

    user foo homedirectory /home/foo is owned by bar

This getpwuid call caused the loop to run forever. We now introduce a
similar call cached_getpwuid that will try to find the user in our
cache. This way we should avoid the user pointer to be reset.