Limit the number of findings per check
Some checks can create a lot of findings (e.g. files without a user, or
without sticky bit set). Creating all findings can consume a lot of
memory and printing thousands of lines in the console is also not really
helpful.

So limit the number of findings per check to 200 and show in the report
if the number of findings is truncated.