Cache shadow entries
Right now we only cache users (pwd) and groups (grp) but no shadow
entries. As a result when we check the password aging for all users we
have to run getspnam for every user. Depending on the number of users
this can be quite expensive.

We now cache all shadow entries in `util.c` like we cache users and
groups. This should allow us to use these cached values in the future.